+++ src/action/check/paymentMethod/movieTicket.ts
@@ -1,18 +1,13 @@
-import { factory as surfrockFactory } from '@surfrock/sdk';
-
 import * as ActionFactory from '../../../action';
 import { OrganizationType } from '../../../organizationType';
 import { IMovieTicketPaymentCard, IMovieTicketServiceOutput } from '../../../paymentMethod/paymentCard/movieTicket';
+import { IMkknInfo, IPurchaseNumberAuthIn, IPurchaseNumberAuthResult, IPurchaseNumberInfo, IRecipe, IYkknInfo } from '../../../recipe/checkMovieTicket';
 import { TransactionType } from '../../../transactionType';
 import * as CheckActionFactory from '../../check';
 import * as PayActionFactory from '../../trade/pay';
 
+export { IMkknInfo, IYkknInfo, IPurchaseNumberInfo, IPurchaseNumberAuthIn, IPurchaseNumberAuthResult, IRecipe };
 export type IAgent = ActionFactory.IParticipantAsWebApplication | ActionFactory.IParticipantAsPerson;
-export type IPurchaseNumberAuthIn = surfrockFactory.service.auth.purchaseNumberAuth.IPurchaseNumberAuthIn;
-export type IPurchaseNumberAuthResult = surfrockFactory.service.auth.purchaseNumberAuth.IPurchaseNumberAuthResult;
-export type IPurchaseNumberInfo = surfrockFactory.service.auth.purchaseNumberAuth.IPurchaseNumberInfo;
-export type IMkknInfo = surfrockFactory.service.auth.purchaseNumberAuth.INvalidTicket;
-export type IYkknInfo = surfrockFactory.service.auth.purchaseNumberAuth.IValidTicket;
 /**
  * 認証対象の決済カード
  */
@@ -47,19 +42,23 @@ export interface IPaymentService extends Pick<
     };
 }
 export type IObject = IPaymentService[];
+// tslint:disable-next-line:no-empty-interface
 export interface IResult {
-    purchaseNumberAuthIn: IPurchaseNumberAuthIn;
-    purchaseNumberAuthResult: IPurchaseNumberAuthResult;
+    // purchaseNumberAuthIn: IPurchaseNumberAuthIn; // discontinue(2024-06-10~)
+    // purchaseNumberAuthResult: IPurchaseNumberAuthResult; // discontinue(2024-06-10~)
 }
 export type IError = any;
 export interface IPurpose {
     typeOf: TransactionType.PlaceOrder;
     id: string;
 }
-export interface IAttributes extends CheckActionFactory.IAttributes<IObject, IResult> {
+export interface IAttributes extends Pick<
+    CheckActionFactory.IAttributes<IObject, IResult>,
+    'agent' | 'error' | 'instrument' | 'object' | 'potentialActions' | 'purpose' | 'result' | 'project' | 'sameAs' | 'typeOf'
+> {
     agent: IAgent;
     object: IObject;
-    purpose?: IPurpose; // add purpose(2023-03-06~)
+    purpose?: IPurpose;
 }
 /**
  * 決済カード認証アクション