+++ src/action/interact/confirm.ts
@@ -9,7 +9,6 @@ import { TransactionType } from '../../transactionType';
  * 決済取引に関してはクライアントによる確定がありうるので拡張(2024-03-11~)
  */
 export type IAgent = ActionFactory.IParticipantAsPerson | ActionFactory.IParticipantAsProject | ActionFactory.IParticipantAsWebApplication;
-// export type IAgent = ActionFactory.IParticipantAsProject;
 export type IObject = any;
 export interface ITransactionPurpose {
     typeOf: TransactionType.MoneyTransfer | TransactionType.PlaceOrder;
@@ -20,7 +19,7 @@ export type IResult = any;
 export type IPotentialActions = any;
 export interface IAttributes<TObject, TResult> extends Pick<
     ActionFactory.IAttributes<ActionType.ConfirmAction, TObject, TResult>,
-    'agent' | 'error' | 'object' | 'project' | 'purpose' | 'result' | 'typeOf'
+    'agent' | 'error' | 'object' | 'project' | 'purpose' | 'result' | 'typeOf' | 'sameAs'
 // | 'potentialActions' // 廃止(2024-03-11~)
 > {
     agent: IAgent;