+++ src/action/interact/confirm/reservation.ts
@@ -1,24 +1,29 @@
-import type * as COA from '@motionpicture/coa-service';
-
 import { IParticipantAsProject } from '../../../action';
 import * as ReserveTransactionFactory from '../../../assetTransaction/reserve';
 import { AssetTransactionType } from '../../../assetTransactionType';
 import { ISimpleOrder } from '../../../order';
+import { IRecipe, IUpdReserveArgs } from '../../../recipe/confirmCOAReserve';
 import * as WebAPIFactory from '../../../service/webAPI';
 import * as ConfirmActionFactory from '../confirm';
 
+export {
+    IRecipe as IConfirmCOAReserveRecipe
+};
 export type IAgent = IParticipantAsProject;
-export type IObject4COA = COA.factory.reserve.IUpdReserveArgs & {
-    // 取引番号は必須
+export type IObject4COA = IUpdReserveArgs & {
+    /**
+     * 取引番号は必須
+     */
     transactionNumber: string;
     typeOf: 'COAReserveTransaction';
 };
 export type IObject4Chevre = Pick<
     ReserveTransactionFactory.IConfirmParams,
-    // 'transactionNumber' | 'object' | 'potentialActions'
     'transactionNumber' | 'potentialActions'
 > & {
-    // 取引番号は必須
+    /**
+     * 取引番号は必須
+     */
     transactionNumber: string;
     typeOf: AssetTransactionType.Reserve;
 };
@@ -27,8 +32,8 @@ export type IObject<T extends WebAPIFactory.Identifier> =
     IObject4Chevre;
 
 export type IPurpose = ISimpleOrder;
-export type IResult = any;
-export type IInstrument<T extends WebAPIFactory.Identifier> = WebAPIFactory.IService<T>;
+// tslint:disable-next-line:no-empty-interface
+export interface IResult { }
 export interface IAttributes<T extends WebAPIFactory.Identifier> extends ConfirmActionFactory.IAttributes<IObject<T>, IResult> {
     agent: IAgent;
     // instrument: IInstrument<T>; // 廃止(2024-03-13~)