+++ src/action/transfer/return/reserveTransaction.ts
@@ -1,14 +1,16 @@
-import type * as COA from '@motionpicture/coa-service';
-
 import * as ActionFactory from '../../../action';
 import { AssetTransactionType } from '../../../assetTransactionType';
 import { ISimpleOrder } from '../../../order';
+import { IRecipe, IStateReserveArgs } from '../../../recipe/returnCOAReserve';
 import { Identifier as WebAPIIdentifier, IService } from '../../../service/webAPI';
 import * as ReturnActionFactory from '../return';
 
+export {
+    IRecipe as IReturnCOAReserveRecipe
+};
 export type IAgent = ActionFactory.IParticipantAsProject;
 export type IRecipient = ActionFactory.IParticipantAsSeller;
-export type IObject4COA = COA.factory.reserve.IStateReserveArgs & {
+export type IObject4COA = IStateReserveArgs & {
     typeOf: 'COAReserveTransaction';
 };
 export interface IObject4Chevre {
@@ -19,9 +21,13 @@ export type IObject<T extends WebAPIIdentifier> =
     T extends WebAPIIdentifier.COA ? IObject4COA :
     IObject4Chevre;
 export type IPurpose = ISimpleOrder;
-export type IResult = any;
-export type IInstrument<T extends WebAPIIdentifier> = IService<T>;
-export interface IAttributes<T extends WebAPIIdentifier> extends ReturnActionFactory.IAttributes<IObject<T>, IResult> {
+// tslint:disable-next-line:no-empty-interface
+export interface IResult { }
+export type IInstrument<T extends WebAPIIdentifier> = Pick<IService<T>, 'identifier' | 'typeOf'>;
+export interface IAttributes<T extends WebAPIIdentifier> extends Pick<
+    ReturnActionFactory.IAttributes<IObject<T>, IResult>,
+    'typeOf' | 'sameAs' | 'result' | 'recipient' | 'purpose' | 'project' | 'object' | 'instrument' | 'error' | 'agent'
+> {
     agent: IAgent;
     instrument: IInstrument<T>;
     purpose: IPurpose;